backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
Authlib==1.6.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
databases==0.9.0
decorator==4.4.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grapheme==0.6.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
moviepy==1.0.3
numpy==2.2.6
oauthlib==3.3.1
passlib==1.7.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
proglog==0.1.12
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pyparsing==3.2.3
pypdfium2==4.30.1
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sarvam==0.0.0
sarvamai==0.1.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
websockets==12.0